+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 0.1.2
+
+* Updated react-sizeme library
+* Add protection for undefined width
+
 ## 0.1.1
 
 * Updated scalajs-react-common